Amman: Added final version of TableQueries

+++ TableQueries.sql
@@ -62,7 +62,7 @@ CREATE TABLE Products
 (
   ProductId SERIAL PRIMARY KEY,
   ProductName VARCHAR(255) NOT NULL,
-  Price INT NOT NULL,
+  Price DECIMAL NOT NULL,
   Inventory INT NOT NULL,
   CategoryId INT NOT NULL,
   SupplierId INT NOT NULL,
@@ -89,7 +89,7 @@ CREATE TABLE OrderDetails
   OrderId INT NOT NULL,
   ProductId INT NOT NULL,
   Quantity INT NOT NULL,
-  TotalPrice INT NOT NULL,
+  TotalPrice DECIMAL NOT NULL,
   FOREIGN KEY(OrderId) REFERENCES Orders(OrderId),
   FOREIGN KEY(ProductId) REFERENCES Products(ProductId)
